requirements.txt
psycopg2-binary
docutils
sentence-transformers
click
